LDrawParser

LDrawParser.BuildingStep

LDrawParser.COMMAND_CODE

LDrawParser.DATModel

LDrawParser.FILE_TYPE

LDrawParser.META_COMMAND

LDrawParser.MPDModel

LDrawParser.NgonElement

LDrawParser.OptionalLineElement

LDrawParser.SubFileRef

LDrawParser.SubModelPlan

LDrawParser.change_coordinate_system!

LDrawParser.ldraw_base_frame

LDrawParser.parse_ldraw_file!

LDrawParser.populate_part_geometry!

LDrawParser.read_line!

LDrawParser.read_meta_line!

LDrawParser.read_optional_line!

LDrawParser.read_quadrilateral!

LDrawParser.read_sub_file_ref!

LDrawParser.read_triangle!

LDrawParser.BuildingStep — Type

BuildingStep

Represents a sequence of part placements that make up a building step in a LDraw file.

LDrawParser.COMMAND_CODE — Type

COMMAND_CODE

The line type of a line is the first number on the line. The line types are:

META # 0 !<META command> <additional parameters>

SUBFILEREF # 1 <colour> x y z a b c d e f g h i <file>

LINE # 2 <colour> x1 y1 z1 x2 y2 z2

TRIANGLE # 3 <colour> x1 y1 z1 x2 y2 z2 x3 y3 z3

QUADRILATERAL # 4 <colour> x1 y1 z1 x2 y2 z2 x3 y3 z3 x4 y4 z4

OPTIONAL_LINE # 5 <colour> x1 y1 z1 x2 y2 z2 x3 y3 z3 x4 y4 z4

LDrawParser.DATModel — Type

DATModel

Encodes the raw geometry of a LDraw part stored in a .dat file. It is possible to avoid populating the geometry fields, which is useful for large models or models that use parts from the LDraw library.

LDrawParser.FILE_TYPE — Type

FILE_TYPE

All LDraw files carry the LDR (default), DAT or MPD extension.

Official Parts Part | Subpart | Primitive | 8Primitive | 48Primitive | Shortcut Unofficial Parts UnofficialPart| UnofficialSubpart | UnofficialPrimitive | Unofficial8Primitive | Unofficial48Primitive | UnofficialShortcut

The file type is usually prefaced in one of the following ways 0 !LDRAWORG <type> (qualifier(s)) (update-tag) 0 LDRAWORG <type> update-tag 0 Official LCAD <type> update-tag 0 Unofficial <type> 0 Un-official <type>

LDrawParser.META_COMMAND — Type

META_COMMAND

0 !<META command> <additional parameters>

! is used to positively identify this as a META command. (Note: A few official meta commands do not start with a ! in order to preserve backwards compatibility, however, all new official META commands must start with a ! and it is strongly recommended that new unofficial meta-commands also start with a !)

<META command> is any string in all caps

<additional parameters> is any string. Note that if a META command does not require any additional parameter, none should be given.

LDrawParser.MPDModel — Type

MPDModel

The MPD model stores the information contained in a .mpd or .ldr file. This includes a submodel tree (stored implicitly in a dictionary that maps model_name to SubModelPlan) and a part list. The first model in MPDModel.models is the main model. All the following are submodels of that model and/or each other.

LDrawParser.NgonElement — Type

NgonElement

Represents geometry from an LDraw file

LDrawParser.OptionalLineElement — Type

OptionalLineElement

Represents optional line geometry from an LDraw file

LDrawParser.SubFileRef — Type

SubFileRef

Represents a sub-file reference from an LDraw file. Encodes the placement of a part or submodel.

LDrawParser.SubModelPlan — Type

SubModelPlan

Represents the sequence of building steps that make up a sub model in an LDraw file

LDrawParser.change_coordinate_system! — Function

change_coordinate_system!(model::MPDModel,T)

Transform the coordinate system of the entire model

LDrawParser.ldraw_base_frame — Method

ldraw_base_frame()

Returns a rotation matrix that defines the base LDraw coordinate system.

LDrawParser.parse_ldraw_file! — Function

parse_ldraw_file!

Args: - model - filename or IO

LDrawParser.populate_part_geometry! — Function

populate_part_geometry!(model,frontier=Set(collect(part_keys(model))))

Load all geometry into model.parts. Loading is recursive, so that geometry will be loaded through arbitrary levels of nested subparts until finally being stored in each atomic part that is referenced by the main model(s).

LDrawParser.read_line! — Method

read_line!

For reading lines of type LINE

LDrawParser.read_meta_line! — Method

read_meta_line(model,state,line)

Modifies the model and parser_state based on a META command. For example, the FILE meta command indicates the beginning of a new file, so this creates a new active model into which subsequent building steps will be placed. The STEP meta command indicates the end of the current step, which prompts the parser to close the current build step and begin a new one.

LDrawParser.read_optional_line! — Method

read_optional_line!

For reading lines of type OPTIONAL_LINE

LDrawParser.read_quadrilateral! — Method

read_quadrilateral!

For reading lines of type QUADRILATERAL

LDrawParser.read_sub_file_ref! — Method

read_sub_file_ref

Receives a SUBFILEREF line (with the leading SUBFILEREF id stripped)

LDrawParser.read_triangle! — Method

read_triangle!

For reading lines of type TRIANGLE